Inline MissionModelFacade outer class
